Update and fix robust GREG predictors
A complete code refactoring of svymean_reg and svytotal_reg (and related functions)

+++ inst/varia/_make_pkg.R
@@ -10,7 +10,7 @@ if (length(args) == 0) {
         tar     = "tar",
         h       = "help",
         help    = "help",
-        pdf     = "pdf")
+        test    =  "test")
 }
 
 if (mode == "help") {
@@ -18,7 +18,8 @@ if (mode == "help") {
     cat("  fast (install only x64, w/o html)\n")
     cat("  check\n")
     cat("  tar (generate tar ball)\n")
-    cat("  full (standard installation)\n\n")
+    cat("  full (standard installation)\n")
+    cat("  test (run test cases)\n\n")
     q(save = "no")
 }
 
@@ -31,30 +32,33 @@ if (.Platform$OS.type == "unix") {
 	PKG_SOURCE <- "C:/My/code"
 	PKG_ROOT <- "C:/My/tmp"
 }
-setwd(PKG_ROOT)
 
-# delete package folder if it already exists
-if (dir.exists(PKG))
-    unlink(PKG, recursive = TRUE)
+if (mode != "test") {
+    setwd(PKG_ROOT)
 
-# copy entire directory (excl. files/folders with a leading dot, e.g. '.git')
-dir.create(PKG)
-pkg_files <- list.files(paste0(PKG_SOURCE, "/", PKG), full.names = TRUE)
-file.copy(pkg_files, paste0(PKG_ROOT, "/", PKG), recursive = TRUE)
+    # delete package folder if it already exists
+    if (dir.exists(PKG))
+        unlink(PKG, recursive = TRUE)
 
-# copy .Rbuildignore and .Rinstignore
-f_R_build_ignore <- paste0(PKG_SOURCE, "/", PKG, "/.Rbuildignore")
-if (file.exists(f_R_build_ignore))
-    file.copy(f_R_build_ignore, paste0(PKG_ROOT, "/", PKG))
-f_R_inst_ignore <- paste0(PKG_SOURCE, "/", PKG, "/.Rinstignore")
-if (file.exists(f_R_inst_ignore))
-    file.copy(f_R_inst_ignore, paste0(PKG_ROOT, "/", PKG))
+    # copy entire directory (excl. files/folders with a leading dot, e.g. '.git')
+    dir.create(PKG)
+    pkg_files <- list.files(paste0(PKG_SOURCE, "/", PKG), full.names = TRUE)
+    file.copy(pkg_files, paste0(PKG_ROOT, "/", PKG), recursive = TRUE)
 
-# clean src folder (remove binary files)
-binary_files <- list.files(paste0(PKG_ROOT, "/", PKG, "/src"),
-    pattern = "\\.o$|\\.dll$|\\.so$")
-if (length(binary_files) > 0)
-    file.remove(paste0(PKG_ROOT, "/", PKG, "/src/", binary_files))
+    # copy .Rbuildignore and .Rinstignore
+    f_R_build_ignore <- paste0(PKG_SOURCE, "/", PKG, "/.Rbuildignore")
+    if (file.exists(f_R_build_ignore))
+        file.copy(f_R_build_ignore, paste0(PKG_ROOT, "/", PKG))
+    f_R_inst_ignore <- paste0(PKG_SOURCE, "/", PKG, "/.Rinstignore")
+    if (file.exists(f_R_inst_ignore))
+        file.copy(f_R_inst_ignore, paste0(PKG_ROOT, "/", PKG))
+
+    # clean src folder (remove binary files)
+    binary_files <- list.files(paste0(PKG_ROOT, "/", PKG, "/src"),
+        pattern = "\\.o$|\\.dll$|\\.so$")
+    if (length(binary_files) > 0)
+        file.remove(paste0(PKG_ROOT, "/", PKG, "/src/", binary_files))
+}
 
 #-------------------------------------------------------------------------------
 setwd(PKG_ROOT)
@@ -88,6 +92,18 @@ if (mode == "full") {
         file.remove(pkg_tar)
     system(paste0("R CMD build ", PKG))
     system(paste0("R CMD INSTALL ", pkg_tar))
+}
+
+# run test cases in folder /test
+if (mode == "test") {
+    setwd(paste0(PKG_SOURCE, "/", PKG, "/tests"))
+    test_cases <- dir(pattern = ".R")
+    cat(rep("-", 40), "\n")
+    for (i in 1:length(test_cases)) {
+        cat(test_cases[i], "\n")
+        system(paste0("Rscript.exe ", test_cases[i]))
+        cat(rep("-", 40), "\n\n")
+    }
 }
 
 # render pdf manual
